middleware applied with some bugs pending to fix

+++ routes/web.php
@@ -26,11 +26,11 @@ Auth::routes();
 Route::get('/',[EventController::class, 'index'])->name('home');
 
 //DELETE
-Route::delete('/delete/{id}', [EventController::class, 'destroy'])->name('delete');
+Route::delete('/delete/{id}', [EventController::class, 'destroy'])->name('delete')->middleware('isadmin', 'auth');
 
 // //CREATE
 
-Route::get('/create',[EventController::class, 'create'])->name('createEvent');
+Route::get('/create',[EventController::class, 'create'])->name('createEvent')->middleware('isadmin', 'auth');
 Route::post('/',[EventController::class, 'store'])->name('storeEvent');
 
 //SHOW
@@ -39,7 +39,6 @@ Route::get('/show/{id}',[EventController::class, 'show'])->name('showEvent');
 
 //UPDATE
 
-Route::get('/edit/{id}', [EventController::class, 'edit'])->name('editEvent');
+Route::get('/edit/{id}', [EventController::class, 'edit'])->name('editEvent')->middleware('isadmin', 'auth');
 ROute::patch('/event/{id}', [EventController::class, 'update'])->name('updateEvent');
 
-HOLAAAAA
